Moving CoreBundle Form Types
From src/CoreBundle/Form to src/CoreBundle/Form/Type
