+++ src/CoreBundle/Form/Type/ActusEditType.php
@@ -1,5 +1,5 @@
 <?php
-namespace CoreBundle\Form;
+namespace CoreBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
