+++ src/CoreBundle/Form/Type/ContactType.php
@@ -1,16 +1,16 @@
 <?php
-namespace CoreBundle\Form;
+namespace CoreBundle\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-use CoreBundle\Form\MyFileType;
-
-class ActusType extends AbstractType
+class ContactType extends AbstractType
 {
 	public function buildForm(
 		FormBuilderInterface $builder,
@@ -19,36 +19,50 @@ class ActusType extends AbstractType
 	{
 		$builder
 			->add(
-				'title',
+				'nom', 
 				TextType::class,
 				array(
-					'label' => 'actus.form.title',
+					'label' => 'index.contact.form.nom',
 					)
-				)
+			)
 			->add(
-				'content',
+				'prenom', 
+				TextType::class,
+				array(
+					'label' => 'index.contact.form.prenom',
+					)
+			)
+			->add(
+				'email', 
+				EmailType::class,
+				array(
+					'label' => 'index.contact.form.email',
+					)
+			)
+			->add(
+				'objet', 
+				TextType::class,
+				array(
+					'label' => 'index.contact.form.objet',
+					)
+			)
+			->add(
+				'contenu', 
 				TextareaType::class,
 				array(
-					'label' => 'actus.form.content',
+					'label' => 'index.contact.form.contenu',
 					)
-				)
+			)
 			->add(
-				'image',
-				MyFileType::class,
-				array(
-					'label' => 'actus.form.image',
-					)
-				)
-			->add(
-				'submit', 
+				'send', 
 				SubmitType::class,
 				array(
-					'label' => 'actus.form.submit',
+					'label' => 'index.contact.form.send',
 					'attr'  => array(
 						'class' => 'btn btn-primary',
 						),
 					)
-				);
+			);
 	}
 
 	public function configureOptions(OptionsResolver $resolver)
@@ -56,7 +70,7 @@ class ActusType extends AbstractType
 		$resolver
 			->setDefaults(
 				array(
-					'data_class' => 'CoreBundle\Entity\Actus'
+					'data_class' => 'CoreBundle\FormModels\ContactModel'
 					)
 				);
 	}
